refactor(constants): rename base to urls, pokemons to pokemon
