+++ src/constants/index.ts
@@ -1,4 +1,3 @@
-export * from "./base";
 export * from "./berries";
 export * from "./contests";
 export * from "./currencies";
@@ -9,5 +8,6 @@ export * from "./games";
 export * from "./items";
 export * from "./locations";
 export * from "./moves";
-export * from "./pokemons";
+export * from "./pokemon";
+export * from "./urls";
 export * from "./utilities";
